Added waveform-view-spec.js for tests common to both views
