+++ test/tests.js
@@ -13,8 +13,9 @@ import './segment-spec';
 import './segment-shape-spec';
 import './segments-layer-spec';
 import './waveform-builder-spec';
-import './waveform-overview-spec';
 import './waveform-shape-spec';
+import './waveform-view-spec';
+import './waveform-overview-spec';
 import './waveform-zoomview-spec';
 import './scrollbar-spec';
 import './keyboard-handler-spec';
