+++ test/waveform-overview-spec.js
@@ -1,107 +1,8 @@
 import Peaks from '../src/main';
-import { Point } from '../src/point';
-import { Segment } from '../src/segment';
 
 import InputController from './helpers/input-controller';
-import { getEmitCalls } from './helpers/utils';
-
-import Konva from 'konva';
 
 describe('WaveformOverview', function() {
-  context('with played waveform color option', function() {
-    it('should create a played waveform shape', function(done) {
-      const options = {
-        overview: {
-          container: document.getElementById('overview-container'),
-          waveformColor: '#f00',
-          playedWaveformColor: '#0f0'
-        },
-        mediaElement: document.getElementById('media'),
-        dataUri: {
-          arraybuffer: 'base/test_data/sample.dat'
-        }
-      };
-
-      Peaks.init(options, function(err, instance) {
-        expect(err).to.equal(null);
-
-        const overview = instance.views.getView('overview');
-        expect(overview).to.be.ok;
-
-        expect(overview._waveformShape._color).to.equal('#f00');
-        expect(overview._playedWaveformShape._color).to.equal('#0f0');
-
-        done();
-      });
-    });
-  });
-
-  describe('setPlayedWaveformColor', function() {
-    it('should create a played waveform shape', function(done) {
-      const options = {
-        overview: {
-          container: document.getElementById('overview-container'),
-          waveformColor: '#f00'
-        },
-        mediaElement: document.getElementById('media'),
-        dataUri: {
-          arraybuffer: 'base/test_data/sample.dat'
-        }
-      };
-
-      Peaks.init(options, function(err, instance) {
-        expect(err).to.equal(null);
-
-        const overview = instance.views.getView('overview');
-        expect(overview).to.be.ok;
-
-        overview.setPlayedWaveformColor('#0f0');
-
-        expect(overview._waveformShape._color).to.equal('#f00');
-        expect(overview._playedWaveformShape._color).to.equal('#0f0');
-
-        done();
-      });
-    });
-
-    it('should remove the played waveform shape', function(done) {
-      const options = {
-        overview: {
-          container: document.getElementById('zoomview-container'),
-          waveformColor: '#f00',
-          playedWaveformColor: '#0f0'
-        },
-        mediaElement: document.getElementById('media'),
-        dataUri: {
-          arraybuffer: 'base/test_data/sample.dat'
-        }
-      };
-
-      Peaks.init(options, function(err, instance) {
-        expect(err).to.equal(null);
-
-        const overview = instance.views.getView('overview');
-        expect(overview).to.be.ok;
-
-        overview.setPlayedWaveformColor(null);
-
-        expect(overview._waveformShape._color).to.equal('#f00');
-        expect(overview._playedWaveformShape).to.equal(null);
-
-        done();
-      });
-    });
-  });
-
-  let p = null;
-
-  afterEach(function() {
-    if (p) {
-      p.destroy();
-      p = null;
-    }
-  });
-
   describe('constructor', function() {
     context('with waveform longer than the container width', function() {
       it('should rescale the waveform to fit the container width', function(done) {
@@ -120,8 +21,6 @@ describe('WaveformOverview', function() {
             done(err);
             return;
           }
-
-          p = instance;
 
           const overview = instance.views.getView('overview');
           expect(overview._data).to.be.ok;
@@ -151,8 +50,6 @@ describe('WaveformOverview', function() {
             return;
           }
 
-          p = instance;
-
           const view = instance.views.getView();
           expect(view._data).to.be.ok;
           expect(view._data.scale).to.equal(32);
@@ -163,6 +60,7 @@ describe('WaveformOverview', function() {
   });
 
   describe('enableSeek', function() {
+    let p = null;
     let inputController = null;
 
     beforeEach(function(done) {
@@ -186,6 +84,13 @@ describe('WaveformOverview', function() {
 
         done();
       });
+    });
+
+    afterEach(function() {
+      if (p) {
+        p.destroy();
+        p = null;
+      }
     });
 
     context('when enabled', function() {
@@ -275,456 +180,6 @@ describe('WaveformOverview', function() {
 
           expect(p.player.getCurrentTime()).to.equal(time);
         });
-      });
-    });
-  });
-
-  describe('click events', function() {
-    let inputController = null;
-
-    beforeEach(function(done) {
-      // TODO: Konva.js uses global state to handle double click timing.
-      // Instead of adding time delays, we just reset Konva's internal
-      // flag here.
-      Konva._mouseInDblClickWindow = false;
-
-      const options = {
-        overview: {
-          container: document.getElementById('overview-container')
-        },
-        mediaElement: document.getElementById('media'),
-        dataUri: { arraybuffer: '/base/test_data/sample.dat' }
-      };
-
-      Peaks.init(options, function(err, instance) {
-        if (err) {
-          done(err);
-          return;
-        }
-
-        p = instance;
-
-        inputController = new InputController('overview-container');
-
-        done();
-      });
-    });
-
-    afterEach(function() {
-      if (p) {
-        p.destroy();
-        p = null;
-      }
-    });
-
-    context('when clicking on the waveform', function() {
-      it('should emit a overview.click event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        inputController.mouseDown({ x: 100, y: 50 });
-        inputController.mouseUp({ x: 100, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview/);
-
-        expect(calls.length).to.equal(1);
-
-        expect(calls[0].args[0]).to.equal('overview.click');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].time).to.be.a('number');
-      });
-
-      it('should emit a overview.dblclick event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        inputController.mouseDown({ x: 100, y: 50 });
-        inputController.mouseUp({ x: 100, y: 50 });
-        inputController.mouseDown({ x: 100, y: 50 });
-        inputController.mouseUp({ x: 100, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview/);
-
-        expect(calls.length).to.equal(3);
-
-        expect(calls[0].args[0]).to.equal('overview.click');
-        expect(calls[1].args[0]).to.equal('overview.click');
-        expect(calls[2].args[0]).to.equal('overview.dblclick');
-
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].time).to.be.a('number');
-      });
-    });
-
-    context('when clicking on a segment', function() {
-      beforeEach(function(done) {
-        p.segments.add({ id: 'segment1', startTime: 1.0, endTime: 2.0, editable: true });
-        setTimeout(done, 50);
-      });
-
-      it('should emit both an overview.click and a segments.click event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        inputController.mouseDown({ x: 40, y: 50 });
-        inputController.mouseUp({ x: 40, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|segments/);
-
-        expect(calls.length).to.equal(4);
-
-        expect(calls[0].args[0]).to.equal('segments.mousedown');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[0].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[1].args[0]).to.equal('segments.mouseup');
-        expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[1].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[1].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[2].args[0]).to.equal('segments.click');
-        expect(calls[2].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[2].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[2].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[3].args[0]).to.equal('overview.click');
-        expect(calls[3].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[3].args[1].time).to.be.a('number');
-      });
-
-      it('should emit both an overview.dblclick and a segments.dblclick event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        inputController.mouseDown({ x: 40, y: 50 });
-        inputController.mouseUp({ x: 40, y: 50 });
-        inputController.mouseDown({ x: 40, y: 50 });
-        inputController.mouseUp({ x: 40, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|segments/);
-
-        expect(calls.length).to.equal(10);
-
-        expect(calls[0].args[0]).to.equal('segments.mousedown');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[0].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[1].args[0]).to.equal('segments.mouseup');
-        expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[1].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[1].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[2].args[0]).to.equal('segments.click');
-        expect(calls[2].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[2].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[2].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[3].args[0]).to.equal('overview.click');
-        expect(calls[3].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[3].args[1].time).to.be.a('number');
-
-        expect(calls[4].args[0]).to.equal('segments.mousedown');
-        expect(calls[4].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[4].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[4].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[5].args[0]).to.equal('segments.mouseup');
-        expect(calls[5].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[5].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[5].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[6].args[0]).to.equal('segments.click');
-        expect(calls[6].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[6].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[6].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[7].args[0]).to.equal('overview.click');
-        expect(calls[7].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[7].args[1].time).to.be.a('number');
-
-        expect(calls[8].args[0]).to.equal('segments.dblclick');
-        expect(calls[8].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[8].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[8].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[9].args[0]).to.equal('overview.dblclick');
-        expect(calls[9].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[9].args[1].time).to.be.a('number');
-      });
-
-      it('should allow the user to prevent the overview.click event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        p.on('segments.click', function(event) {
-          event.preventViewEvent();
-        });
-
-        inputController.mouseDown({ x: 40, y: 50 });
-        inputController.mouseUp({ x: 40, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|segments/);
-
-        expect(calls.length).to.equal(3);
-
-        expect(calls[0].args[0]).to.equal('segments.mousedown');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[0].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[1].args[0]).to.equal('segments.mouseup');
-        expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[1].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[1].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[2].args[0]).to.equal('segments.click');
-        expect(calls[2].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[2].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[2].args[1].segment.id).to.equal('segment1');
-      });
-
-      it('should allow the user to prevent the overview.dblclick event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        p.on('segments.dblclick', function(event) {
-          event.preventViewEvent();
-        });
-
-        inputController.mouseDown({ x: 40, y: 50 });
-        inputController.mouseUp({ x: 40, y: 50 });
-        inputController.mouseDown({ x: 40, y: 50 });
-        inputController.mouseUp({ x: 40, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|segments/);
-
-        expect(calls.length).to.equal(9);
-
-        expect(calls[0].args[0]).to.equal('segments.mousedown');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[0].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[1].args[0]).to.equal('segments.mouseup');
-        expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[1].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[1].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[2].args[0]).to.equal('segments.click');
-        expect(calls[2].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[2].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[2].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[3].args[0]).to.equal('overview.click');
-        expect(calls[3].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[3].args[1].time).to.be.a('number');
-
-        expect(calls[4].args[0]).to.equal('segments.mousedown');
-        expect(calls[4].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[4].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[4].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[5].args[0]).to.equal('segments.mouseup');
-        expect(calls[5].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[5].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[5].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[6].args[0]).to.equal('segments.click');
-        expect(calls[6].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[6].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[6].args[1].segment.id).to.equal('segment1');
-
-        expect(calls[7].args[0]).to.equal('overview.click');
-        expect(calls[7].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[7].args[1].time).to.be.a('number');
-
-        expect(calls[8].args[0]).to.equal('segments.dblclick');
-        expect(calls[8].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[8].args[1].segment).to.be.an.instanceOf(Segment);
-        expect(calls[8].args[1].segment.id).to.equal('segment1');
-      });
-    });
-
-    context('when clicking on a point', function() {
-      beforeEach(function(done) {
-        p.points.add({ id: 'point1', time: 1.0, editable: true });
-        setTimeout(done, 50);
-      });
-
-      it('should emit both an overview.click and a points.click event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        inputController.mouseDown({ x: 30, y: 50 });
-        inputController.mouseUp({ x: 30, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|points/);
-
-        expect(calls.length).to.equal(2);
-
-        // expect(calls[0].args[0]).to.equal('points.mousedown');
-        // expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[0].args[1].point.id).to.equal('point1');
-
-        // expect(calls[1].args[0]).to.equal('points.mouseup');
-        // expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[1].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[1].args[1].point.id).to.equal('point1');
-
-        expect(calls[0].args[0]).to.equal('points.click');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[0].args[1].point.id).to.equal('point1');
-
-        expect(calls[1].args[0]).to.equal('overview.click');
-        expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[1].args[1].time).to.be.a('number');
-      });
-
-      it('should emit both an overview.dblclick and a points.dblclick event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        inputController.mouseDown({ x: 30, y: 50 });
-        inputController.mouseUp({ x: 30, y: 50 });
-        inputController.mouseDown({ x: 30, y: 50 });
-        inputController.mouseUp({ x: 30, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|points/);
-
-        expect(calls.length).to.equal(6);
-
-        // expect(calls[0].args[0]).to.equal('points.mousedown');
-        // expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[0].args[1].point.id).to.equal('segment1');
-
-        // expect(calls[1].args[0]).to.equal('points.mouseup');
-        // expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[1].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[1].args[1].point.id).to.equal('segment1');
-
-        expect(calls[0].args[0]).to.equal('points.click');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[0].args[1].point.id).to.equal('point1');
-
-        expect(calls[1].args[0]).to.equal('overview.click');
-        expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[1].args[1].time).to.be.a('number');
-
-        // expect(calls[4].args[0]).to.equal('points.mousedown');
-        // expect(calls[4].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[4].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[4].args[1].point.id).to.equal('point1');
-
-        // expect(calls[5].args[0]).to.equal('points.mouseup');
-        // expect(calls[5].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[5].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[5].args[1].point.id).to.equal('point1');
-
-        expect(calls[2].args[0]).to.equal('points.click');
-        expect(calls[2].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[2].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[2].args[1].point.id).to.equal('point1');
-
-        expect(calls[3].args[0]).to.equal('overview.click');
-        expect(calls[3].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[3].args[1].time).to.be.a('number');
-
-        expect(calls[4].args[0]).to.equal('points.dblclick');
-        expect(calls[4].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[4].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[4].args[1].point.id).to.equal('point1');
-
-        expect(calls[5].args[0]).to.equal('overview.dblclick');
-        expect(calls[5].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[5].args[1].time).to.be.a('number');
-      });
-
-      it('should allow the user to prevent the overview.click event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        p.on('points.click', function(event) {
-          event.preventViewEvent();
-        });
-
-        inputController.mouseDown({ x: 30, y: 50 });
-        inputController.mouseUp({ x: 30, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|points/);
-
-        expect(calls.length).to.equal(1);
-
-        // expect(calls[0].args[0]).to.equal('points.mousedown');
-        // expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[0].args[1].point.id).to.equal('point1');
-
-        // expect(calls[1].args[0]).to.equal('points.mouseup');
-        // expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[1].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[1].args[1].point.id).to.equal('point1');
-
-        expect(calls[0].args[0]).to.equal('points.click');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[0].args[1].point.id).to.equal('point1');
-      });
-
-      it('should allow the user to prevent the overview.dblclick event', function() {
-        const emit = sinon.spy(p, 'emit');
-
-        p.on('points.dblclick', function(event) {
-          event.preventViewEvent();
-        });
-
-        inputController.mouseDown({ x: 30, y: 50 });
-        inputController.mouseUp({ x: 30, y: 50 });
-        inputController.mouseDown({ x: 30, y: 50 });
-        inputController.mouseUp({ x: 30, y: 50 });
-
-        const calls = getEmitCalls(emit, /overview|points/);
-
-        expect(calls.length).to.equal(5);
-
-        // expect(calls[0].args[0]).to.equal('points.mousedown');
-        // expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[0].args[1].point.id).to.equal('point1');
-
-        // expect(calls[1].args[0]).to.equal('points.mouseup');
-        // expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[1].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[1].args[1].point.id).to.equal('point1');
-
-        expect(calls[0].args[0]).to.equal('points.click');
-        expect(calls[0].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[0].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[0].args[1].point.id).to.equal('point1');
-
-        expect(calls[1].args[0]).to.equal('overview.click');
-        expect(calls[1].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[1].args[1].time).to.be.a('number');
-
-        // expect(calls[4].args[0]).to.equal('points.mousedown');
-        // expect(calls[4].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[4].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[4].args[1].point.id).to.equal('point1');
-
-        // expect(calls[5].args[0]).to.equal('points.mouseup');
-        // expect(calls[5].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        // expect(calls[5].args[1].point).to.be.an.instanceOf(Point);
-        // expect(calls[5].args[1].point.id).to.equal('point1');
-
-        expect(calls[2].args[0]).to.equal('points.click');
-        expect(calls[2].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[2].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[2].args[1].point.id).to.equal('point1');
-
-        expect(calls[3].args[0]).to.equal('overview.click');
-        expect(calls[3].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[3].args[1].time).to.be.a('number');
-
-        expect(calls[4].args[0]).to.equal('points.dblclick');
-        expect(calls[4].args[1].evt).to.be.an.instanceOf(MouseEvent);
-        expect(calls[4].args[1].point).to.be.an.instanceOf(Point);
-        expect(calls[4].args[1].point.id).to.equal('point1');
       });
     });
   });
